+++ seqkit/main.go
@@ -1,4 +1,4 @@
-// Copyright © 2016 Wei Shen <shenwei356@gmail.com>
+// Copyright © 2016-2019 Wei Shen <shenwei356@gmail.com>
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -56,6 +56,7 @@ func main() {
 	cmd.Execute()
 
 	// go tool pprof --alloc_space /usr/local/bin/seqkit mprof
+	// go tool pprof --inuse_space /usr/local/bin/seqkit mprof
 	// f2, _ := os.Create("mprof")
 	// pprof.WriteHeapProfile(f2)
 	// defer f2.Close()